@@ -1,6 +1,6 @@
 module github.com/kyma-project/kyma-metrics-collector
 
-go 1.23.6
+go 1.24.0
 
 require (
 	github.com/avast/retry-go/v4 v4.6.0
