@@ -1,7 +1,6 @@
 package redis
 
 import (
-	"context"
 	"errors"
 	"testing"
 
@@ -52,7 +51,7 @@ func TestScanner_Scan_Successful(t *testing.T) {
 
 	provider := "test-provider"
 
-	result, err := scanner.Scan(context.Background(), &runtime.Info{
+	result, err := scanner.Scan(t.Context(), &runtime.Info{
 		ProviderType: provider,
 	}, clients)
 	require.NoError(t, err)
@@ -80,7 +79,7 @@ func TestScanner_Scan_Error(t *testing.T) {
 
 	scanner := Scanner{}
 
-	result, err := scanner.Scan(context.Background(), &runtime.Info{}, clients)
+	result, err := scanner.Scan(t.Context(), &runtime.Info{}, clients)
 
 	require.Error(t, err)
 	require.Nil(t, result)
